Moved incrementer into dedicated file
