+++ protocol/anonmetrics/client.go
@@ -85,17 +85,3 @@ func (c *Client) Stop() error {
 	c.stopMainLoop()
 	return nil
 }
-
-type FibonacciIntervalIncrementer struct {
-	Last    int64
-	Current int64
-}
-
-func (f *FibonacciIntervalIncrementer) Next() int64 {
-	out := f.Last + f.Current
-
-	f.Last = f.Current
-	f.Current = out
-
-	return out
-}
